requirements.txt
torch
torchvision
torchaudio
xformers
webdataset
wandb
tqdm
pyyaml
matplotlib
pillow
